Split lib.py into util Package
