+++ package_install.py
@@ -1,5 +1,5 @@
 #!/bin/python3
-from lib import Print, Install
+from util import Print, Install
 
 
 def main():
@@ -9,15 +9,20 @@ def main():
 
     Install.package('networkmanager')
     Install.package(
-        'git', 'default + (Required for some installations, otherwise they\'ll be skipped)')
+        'git',
+        'default + (Required for some installations, otherwise they\'ll be skipped)'
+    )
     Install.package('sudo')
 
     # Desktop Enviroment
-    Install.multiple_packages(['plasma', 'plasma-desktop', 'gnome'], 'Do you wish to install DE (Desktop Enviroment)?', [
-        'Plasma (KDE)', 'Plasma-Desktop (KDE-Minimal dependencies)', 'Gnome'])
+    Install.multiple_packages(
+        ['plasma', 'plasma-desktop', 'gnome'],
+        'Do you wish to install DE (Desktop Enviroment)?',
+        ['Plasma (KDE)', 'Plasma-Desktop (KDE-Minimal dependencies)', 'Gnome'])
     # Display Manager
-    Install.multiple_packages(['sddm', 'gdm', 'lightdm'], 'Do you wish to install DM (Display Manager)?', [
-        'SDDM (KDE)', 'GDM (Gnome)', 'LightDM'])
+    Install.multiple_packages(['sddm', 'gdm', 'lightdm'],
+                              'Do you wish to install DM (Display Manager)?',
+                              ['SDDM (KDE)', 'GDM (Gnome)', 'LightDM'])
 
     Install.package('base-devel', 'default + (Required for makepkg installs)')
     Install.package('yay', 'default', aur=True)
@@ -28,25 +33,27 @@ def main():
     Install.package('terminator', 'default + (advanced terminal)')
     Install.package('konsole', 'default + (KDE terminal emulator)')
     Install.package(
-        'ark', 'default + (Managing various archive formats such as tar, gzip, zip, rar, etc.)')
+        'ark',
+        'default + (Managing various archive formats such as tar, gzip, zip, rar, etc.)'
+    )
     Install.package('timeshift', 'default + (Backup Tool), aur=True')
     Install.package('cron', 'default + (Task scheduling)')
     Install.package('dolphin', 'default + (File Manager)')
     Install.package('nomacs', 'default + (Photo viewer & editor)')
     Install.package('discord', 'default + (Chat App)')
     Install.package('caprine', 'default + (Unofficial Messenger Chat App)')
-    Install.package(
-        'spotify', 'default + (Online Music Player)', aur=True)
-    Install.package(
-        'spectacle', 'default + (Screenshot tool)')
+    Install.package('spotify', 'default + (Online Music Player)', aur=True)
+    Install.package('spectacle', 'default + (Screenshot tool)')
     Install.package('qalculate-gtk-nognome',
-                    'Do you wish to install Qalculate! (Complex Calculator)?', aur=True)
+                    'Do you wish to install Qalculate! (Complex Calculator)?',
+                    aur=True)
     Install.package('gnome-system-monitor',
                     'Do you wish to install gnome system monitor?')
     Install.package(
-        'code', 'Do you wish to install Visual Studio Code (Text/Code Editor)?')
-    Install.package(
-        'filelight', 'default + (Disk usage statistics and graphs)')
+        'code',
+        'Do you wish to install Visual Studio Code (Text/Code Editor)?')
+    Install.package('filelight',
+                    'default + (Disk usage statistics and graphs)')
     Install.multiple_packages(['firefox', 'chromium'],
                               'Do you wish to install web browser?')
